calcSOM

calculates Soil Organic Matter Pool, accounting for the management history. We assume carbon Stocks from LPJml natural vegetation as a starting point. Here we use the upper 30cm soil layer (0-20cm of + 1/3 of 30-50 cm). We then correct carbon pools by lost c-share depending on the climate region, using default factors of IPCC Guidelines 2006 table 5.5. We assume that this IPCC-corrected value is the target long-term equilibrium value for the soil stocks. Because soil decline and build-up slowly, we assume that in every year, the carbon pools move 15% towards this new equilibrium. This assumption is in line with IPCC saying that the process will take 20 years: with our assumption, after 5 years 44% of the carbon pool is gone, after 10 years 80% and after 20 years 96%. We determine a carbon stock for cropland soils and non-cropland soils in every cell. If the cropland area expands, the carbon stock of noncropland is proportionally assigned to the cropland pool and vice versa. The outputs of the function are the soilc stocks for cropland and non-cropland. Relevant for the release of N by SOM loss is also the change in carbon stocks per ha, as this relases or binds N. This is done in delta cropland soilc.

Usage

calcSOM(climatetype = "historical", subtype = "stock", cells = "lpjcell")

Arguments

climatetype

Switch between different climate scenarios (default on "historical")

subtype

"stock" (default) for absoulte values, "density" for per hectar values

cells

(deprecated) Only option "lpjcell" for 67420 cells

Value

List of magpie object with results on country or cellular level, weight on cellular level, unit and description.

Author

Benjamin Leon Bodirsky, Kristine Karstens

Examples

if (FALSE) { # \dontrun{
calcOutput("SOM2")
} # }